feat(containers): Miscellaneous container fixes

+++ containers/sandbox-with-ghcr.bash
@@ -21,12 +21,15 @@ EOF
 
 sandbox() {
 	local -r tag="$1"
-	local -r full_image_id="$(podman pull "ghcr.io/kachick/home:${tag}")"
+	local -r full_image_id="$(podman pull "ghcr.io/joaqim/home:${tag}")"
+	if [ -z "$full_image_id" ]; then
+		exit 1
+	fi
 	local -r container_id="$(podman run --detach --rm "$full_image_id")"
 	# shellcheck disable=SC2064
 	trap "podman kill '$container_id'" EXIT ERR
 	sleep 1 # Wait for the systemd to be ready
-	podman exec --user=user --workdir='/home/user' -eTERM -it "$container_id" '/home/user/.nix-profile/bin/zsh'
+	podman exec --user=user --workdir='/home/user' -eTERM -it "$container_id" '/home/user/.nix-profile/bin/nu'
 }
 
 sandbox "$1"
